Add documentation to testsuite
Added initial documentation for testsuite. Included ADR, architecture,
configuration, deployment, development, index, secrets, system
landscape and tests.

Described physical setup in architecture, as it is impossible to
figure it out by yourself. Described the first part of a testsuite and
touched on the scope of the testsuite and what it should be able to cover.